Update map center and bonds after route calculation
